requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
asgiref==3.10.0
attrs==25.4.0
boto3==1.40.65
botocore==1.40.65
BTrees==6.2
certifi==2025.10.5
cffi==2.0.0
Chameleon==4.6.0
charset-normalizer==3.4.4
cloudinary==1.44.1
crispy-bootstrap4==2025.6
Django==5.0
django-allauth==65.12.0
django-cloudinary-storage==0.3.0
django-colorfield==0.14.0
django-countries==7.6.1
django-crispy-forms==2.4
django-phonenumber-field==8.3.0
django-ratelimit==4.1.0
django-redis==6.0.0
django-storages==1.14.6
django-tenants==3.9.0
django-twilio==0.14.3.3
django-widget-tweaks==1.5.0
frozenlist==1.8.0
grok==6.1
grokcore.annotation==5.0
grokcore.catalog==5.0
grokcore.chameleon==5.0
grokcore.component==5.0
grokcore.content==5.0
grokcore.formlib==5.0
grokcore.layout==5.0
grokcore.message==5.0
grokcore.security==5.0
grokcore.site==5.0
grokcore.traverser==5.0
grokcore.view==5.0
grokcore.viewlet==5.0
idna==3.11
jmespath==1.0.1
legacy-cgi==2.6.4
martian==2.1
multidict==6.7.0
multipart==1.3.0
persistent==6.3
phonenumbers==9.0.16
pillow==11.3.0
propcache==0.4.1
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-gettext==5.0
pytz==2025.2
redis==7.0.1
requests==2.32.5
s3transfer==0.14.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
transaction==5.0
twilio==9.8.5
typing_extensions==4.15.0
urllib3==2.5.0
yarl==1.22.0
z3c.flashmessage==4.0
z3c.pt==5.1
zc.catalog==5.0
zc.lockfile==4.0
ZConfig==4.2
zdaemon==5.2.1
ZODB==6.1
zodbpickle==4.2
zope.annotation==6.0
zope.app.appsetup==6.0
zope.app.publication==6.0
zope.app.wsgi==6.0
zope.authentication==6.0
zope.browser==4.0
zope.browserpage==6.0
zope.browserresource==6.0
zope.cachedescriptors==6.0
zope.catalog==6.0
zope.component==7.0
zope.configuration==7.0
zope.container==7.0
zope.contentprovider==7.0
zope.contenttype==6.0
zope.datetime==6.0
zope.deferredimport==6.0
zope.deprecation==6.0
zope.dottedname==7.0
zope.error==6.0
zope.errorview==3.0
zope.event==6.0
zope.exceptions==6.0
zope.filerepresentation==7.0
zope.formlib==7.0
zope.generations==7.0
zope.hookable==8.1
zope.i18n==6.0
zope.i18nmessageid==8.1
zope.index==8.0
zope.interface==8.0.1
zope.intid==6.0
zope.keyreference==7.0
zope.lifecycleevent==6.0
zope.location==6.0
zope.login==4.0
zope.minmax==4.0
zope.pagetemplate==6.0
zope.password==6.1
zope.principalregistry==6.0
zope.processlifetime==4.0
zope.proxy==7.0
zope.ptresource==6.0
zope.publisher==8.0
zope.schema==8.1
zope.security==8.2
zope.securitypolicy==6.0
zope.session==6.0
zope.site==6.0
zope.size==6.0
zope.tal==6.0
zope.tales==7.0
zope.testing==6.0
zope.traversing==6.0
zope.viewlet==6.0
